pip-requirements.txt
# Install these requirements with pip install -r pip-requirements.txt. 
ckanext-canada
geojson==1.0.1
lxml==3.1.0
matplotlib==1.2.0
multi-mechanize==1.2.0
numpy==1.7.0
pysqlite==2.6.3
simplejson==3.1.0
wsgiref==0.1.2
guess-language
